svg/painting/parsing/stroke-linecap-valid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing stroke-linecap with valid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#StrokeLinecapProperty" rel="help"/><h:meta content="stroke-linecap supports the full grammar 'butt | round | square'" name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_valid_value("stroke-linecap", "butt");
test_valid_value("stroke-linecap", "round");
test_valid_value("stroke-linecap", "square");

  ]]></script></svg>
